Make the interactive plot scripts create and check for a lock file before plotting.

+++ source/plotbridge/plotbridge.py
@@ -155,6 +155,7 @@ class Plot():
       logging.warn('Removing contents of old "%s"', d)
       try:
         for f in os.listdir(d):
+          if f.endswith('.lock'): continue # don't remove lock files
           ppp = os.path.join(d,f)
           if   os.path.isfile(ppp) or os.path.islink(ppp): os.unlink(ppp)
           elif os.path.isdir(ppp): logging.warn('Ignoring directory %s', ppp) #shutil.rmtree(ppp)  # Normally there are no subdirs, so stay on the safe side and don't delete them...
